+++ NCexamples/qtransport/Makefile
@@ -1,0 +1,2 @@
+clean:
+	rm -r 1elec 2channel 3postprocess *fdf
