--- NCexamples/qtransport/transmission/1.elec/Au_left/input/STRUCT.fdf
@@ -1,50 +1,0 @@
-NumberOfAtoms    36           # Number of atoms
-NumberOfSpecies  1           # Number of species
-%block ChemicalSpeciesLabel
- 1 79 Au
-%endblock ChemicalSpeciesLabel
-LatticeConstant   1.0 Ang
-%block LatticeVectors
-    8.472000000    -0.000000000    -0.000000000
-    0.000000000     9.782620000    -0.000000000
-    0.000000000     0.000000000     6.917360000
-%endblock LatticeVectors
-AtomicCoordinatesFormat Ang
-%block AtomicCoordinatesAndAtomicSpecies
-     2.118000000     2.038046000     1.152893000    1    1
-     0.706000000     4.483702000     1.152893000    1    2
-     4.942000000     2.038046000     1.152893000    1    3
-     3.530000000     4.483702000     1.152893000    1    4
-     7.766000000     2.038046000     1.152893000    1    5
-     6.354000000     4.483702000     1.152893000    1    6
-     2.118000000     6.929358000     1.152893000    1    7
-     0.706000000     9.375014000     1.152893000    1    8
-     4.942000000     6.929358000     1.152893000    1    9
-     3.530000000     9.375014000     1.152893000    1   10
-     7.766000000     6.929358000     1.152893000    1   11
-     6.354000000     9.375014000     1.152893000    1   12
-     2.118000000     0.407609000     3.458679000    1   13
-     0.706000000     2.853265000     3.458679000    1   14
-     4.942000000     0.407609000     3.458679000    1   15
-     3.530000000     2.853265000     3.458679000    1   16
-     7.766000000     0.407609000     3.458679000    1   17
-     6.354000000     2.853265000     3.458679000    1   18
-     2.118000000     5.298921000     3.458679000    1   19
-     0.706000000     7.744577000     3.458679000    1   20
-     4.942000000     5.298921000     3.458679000    1   21
-     3.530000000     7.744577000     3.458679000    1   22
-     7.766000000     5.298921000     3.458679000    1   23
-     6.354000000     7.744577000     3.458679000    1   24
-     0.706000000     1.222828000     5.764465000    1   25
-     2.118000000     3.668484000     5.764465000    1   26
-     3.530000000     1.222828000     5.764465000    1   27
-     4.942000000     3.668484000     5.764465000    1   28
-     6.354000000     1.222828000     5.764465000    1   29
-     7.766000000     3.668484000     5.764465000    1   30
-     0.706000000     6.114139000     5.764465000    1   31
-     2.118000000     8.559795000     5.764465000    1   32
-     3.530000000     6.114139000     5.764465000    1   33
-     4.942000000     8.559795000     5.764465000    1   34
-     6.354000000     6.114139000     5.764465000    1   35
-     7.766000000     8.559795000     5.764465000    1   36
-%endblock AtomicCoordinatesAndAtomicSpecies
