--- NCexamples/qtransport/transmission/1.elec/Au_right/input/STRUCT.fdf
@@ -1,50 +1,0 @@
-NumberOfAtoms    36           # Number of atoms
-NumberOfSpecies  1           # Number of species
-%block ChemicalSpeciesLabel
- 1 79 Au
-%endblock ChemicalSpeciesLabel
-LatticeConstant   1.0 Ang
-%block LatticeVectors
-    8.472000000    -0.000000000    -0.000000000
-    0.000000000     9.782620000    -0.000000000
-    0.000000000     0.000000000     6.917360000
-%endblock LatticeVectors
-AtomicCoordinatesFormat Ang
-%block AtomicCoordinatesAndAtomicSpecies
-     0.706000000     1.222828000     1.152893000    1    1
-     2.118000000     3.668484000     1.152893000    1    2
-     3.530000000     1.222828000     1.152893000    1    3
-     4.942000000     3.668484000     1.152893000    1    4
-     6.354000000     1.222828000     1.152893000    1    5
-     7.766000000     3.668484000     1.152893000    1    6
-     0.706000000     6.114139000     1.152893000    1    7
-     2.118000000     8.559795000     1.152893000    1    8
-     3.530000000     6.114139000     1.152893000    1    9
-     4.942000000     8.559795000     1.152893000    1   10
-     6.354000000     6.114139000     1.152893000    1   11
-     7.766000000     8.559795000     1.152893000    1   12
-     2.118000000     0.407609000     3.458679000    1   13
-     0.706000000     2.853265000     3.458679000    1   14
-     4.942000000     0.407609000     3.458679000    1   15
-     3.530000000     2.853265000     3.458679000    1   16
-     7.766000000     0.407609000     3.458679000    1   17
-     6.354000000     2.853265000     3.458679000    1   18
-     2.118000000     5.298921000     3.458679000    1   19
-     0.706000000     7.744577000     3.458679000    1   20
-     4.942000000     5.298921000     3.458679000    1   21
-     3.530000000     7.744577000     3.458679000    1   22
-     7.766000000     5.298921000     3.458679000    1   23
-     6.354000000     7.744577000     3.458679000    1   24
-     2.118000000     2.038046000     5.764465000    1   25
-     0.706000000     4.483702000     5.764465000    1   26
-     4.942000000     2.038046000     5.764465000    1   27
-     3.530000000     4.483702000     5.764465000    1   28
-     7.766000000     2.038046000     5.764465000    1   29
-     6.354000000     4.483702000     5.764465000    1   30
-     2.118000000     6.929358000     5.764465000    1   31
-     0.706000000     9.375014000     5.764465000    1   32
-     4.942000000     6.929358000     5.764465000    1   33
-     3.530000000     9.375014000     5.764465000    1   34
-     7.766000000     6.929358000     5.764465000    1   35
-     6.354000000     9.375014000     5.764465000    1   36
-%endblock AtomicCoordinatesAndAtomicSpecies
