--- NCexamples/qtransport/transmission/2.model/initial/elecR.fdf
@@ -1,67 +1,0 @@
-NumberOfAtoms    48           # Number of atoms
-NumberOfSpecies  1           # Number of species
-
-%block ChemicalSpeciesLabel
- 1 79 Au
-%endblock ChemicalSpeciesLabel
-
-#(3) Lattice, coordinates, k-sampling
-
-LatticeConstant       1.000000000 Ang
-%block LatticeVectors
-    8.472000000     0.000000000     0.000000000
-    0.000000000     9.782620000     0.000000000
-    0.000000000     0.000000000    42.298530000
-%endblock LatticeVectors
-
-AtomicCoordinatesFormat Ang
-%block AtomicCoordinatesAndAtomicSpecies
-     2.118000000     2.038046000    41.145641000    1    1
-     0.706000000     4.483702000    41.145641000    1    2
-     4.942000000     2.038046000    41.145641000    1    3
-     3.530000000     4.483702000    41.145641000    1    4
-     7.766000000     2.038046000    41.145641000    1    5
-     6.354000000     4.483702000    41.145641000    1    6
-     2.118000000     6.929358000    41.145641000    1    7
-     0.706000000     9.375014000    41.145641000    1    8
-     4.942000000     6.929358000    41.145641000    1    9
-     3.530000000     9.375014000    41.145641000    1   10
-     7.766000000     6.929358000    41.145641000    1   11
-     6.354000000     9.375014000    41.145641000    1   12
-     2.118000000     0.407609000    38.839855000    1   13
-     0.706000000     2.853265000    38.839855000    1   14
-     4.942000000     0.407609000    38.839855000    1   15
-     3.530000000     2.853265000    38.839855000    1   16
-     7.766000000     0.407609000    38.839855000    1   17
-     6.354000000     2.853265000    38.839855000    1   18
-     2.118000000     5.298921000    38.839855000    1   19
-     0.706000000     7.744577000    38.839855000    1   20
-     4.942000000     5.298921000    38.839855000    1   21
-     3.530000000     7.744577000    38.839855000    1   22
-     7.766000000     5.298921000    38.839855000    1   23
-     6.354000000     7.744577000    38.839855000    1   24
-     0.706000000     1.222828000    36.534069000    1   25
-     2.118000000     3.668484000    36.534069000    1   26
-     3.530000000     1.222828000    36.534069000    1   27
-     4.942000000     3.668484000    36.534069000    1   28
-     6.354000000     1.222828000    36.534069000    1   29
-     7.766000000     3.668484000    36.534069000    1   30
-     0.706000000     6.114139000    36.534069000    1   31
-     2.118000000     8.559795000    36.534069000    1   32
-     3.530000000     6.114139000    36.534069000    1   33
-     4.942000000     8.559795000    36.534069000    1   34
-     6.354000000     6.114139000    36.534069000    1   35
-     7.766000000     8.559795000    36.534069000    1   36
-     2.118000000     2.038046000    34.228282000    1   37
-     0.706000000     4.483702000    34.228282000    1   38
-     4.942000000     2.038046000    34.228282000    1   39
-     3.530000000     4.483702000    34.228282000    1   40
-     7.766000000     2.038046000    34.228282000    1   41
-     6.354000000     4.483702000    34.228282000    1   42
-     2.118000000     6.929358000    34.228282000    1   43
-     0.706000000     9.375014000    34.228282000    1   44
-     4.942000000     6.929358000    34.228282000    1   45
-     3.530000000     9.375014000    34.228282000    1   46
-     7.766000000     6.929358000    34.228282000    1   47
-     6.354000000     9.375014000    34.228282000    1   48
-%endblock AtomicCoordinatesAndAtomicSpecies
