--- NCexamples/qtransport/transmission/2.model/initial/mirror.py
@@ -1,5 +1,0 @@
-import nanocore.io as nc
-
-atms = nc.read_struct('elecL.fdf')
-atms.get_mirrored_structure()
-nc.write_struct(atms)
