--- NCexamples/qtransport/transmission/scripts/Makefile
@@ -1,19 +1,0 @@
-cp:
-	cp *py Makefile readme.txt scripts/
-
-cp2nc:
-	cp scripts/* $${NCHOME}/NCexamples/qtransport/transmission/scripts
-
-
-clean3:
-	rm -r   3.scatter_tbtrans/0*
-
-clean4:
-	rm -r 	4.post_processing/files 4.post_processing/result* 4.post_processing/output.yaml
-
-clean:
-	rm      lap*
-	rm -r 	1.elec/Au*/OUT 
-	rm      2.model/*fdf
-	rm -r 	3.scatter_tbtrans/0*
-	rm -r 	4.post_processing/files 4.post_processing/result* 4.post_processing/output.yaml
