+++ README
@@ -1,11 +1,25 @@
 KAIST EE YHKLab 
 NanoCore in development
 ============================= developed by J Park ==========================================================
-# download
+###### Download and setting
+### Download
 git clone https://github.com/yhkimlab/NanoCore nanocore
 
-# set PYTHONPATH to the front of nanocore or mv nanocore to any $PYTHONPATH such as .../site-packages; e.g.
+### setting
+#set PYTHONPATH to the front of nanocore or mv nanocore to any $PYTHONPATH such as .../site-packages; e.g.
 export PYTHONPATH=$path_to_before_nanocore:$PYTHONPATH
+#set PATH to nanocore/bin or mv nanocore/bin/* to your bin directory
+export PATH=$path_to_nanocre/bin
 
-# copy example directory and run examples at $work_dir
+#modify aux/env.py for pathway to simulator
+
+###### Copy examples and refer to readme file in each example directory
+### copy example directory and run examples at $your_work_dir
 cp -r NCexamples $work_dir
+
+### or partly copy present working examples
+cp -r NCexamples/catalysis $your_work_dir
+cp -r NCexamples/qtransport $your_work_dir
+
+
+
