+++ simulator/siesta/__init__.py
@@ -1,9 +1,8 @@
 #from __future__ import print_function
 from ...atoms import *
-#from ...ncio import cleansymb, get_unique_symbs, convert_xyz2abc, write_xsf
-from ...ncio import write_struct, cleansymb, get_unique_symbs, ang2bohr, read_struct
-from ...aux import convert_xyz2abc 
-from ...units import ang2bohr
+from ...ncio import cleansymb, get_unique_symbs
+from ...aux import convert_abc2xyz, convert_xyz2abc, read_xyz
+from ...units import ang2bohr,  degrad, bohr2ang
 from glob import glob
 import re, sys
 import shutil
@@ -11,6 +10,7 @@ import subprocess
 ### inside siesta_defalut_location, psf, model, elec exist
 from ...aux.env import siesta_default_location as default_location
 
+from math import sin, cos, sqrt, pi
 
 #from ...aux.env import siesta_dir as sul
 #from ...aux.env import siesta_util_vh as sv
@@ -452,7 +452,7 @@ class Siesta(object):
         self._inputs[fname] = lines
         return lines
 
-    def run_qt(self, nproc, **option):
+    def runQuantumTransport(self, nproc, **option):
 
         """
         Run a quantum transport simulation based on the information saved in this simulation object
@@ -607,6 +607,203 @@ class Siesta(object):
                      np.array(self._atoms.get_cell())], 
                     open(name,'w'))
         print ("simulation information is saved as %s." % name)
+
+###### Functional in siesta module
+### write and read structure in fdf format
+def write_struct(atoms, cellparameter=1.0, fname = "STRUCT.fdf"):
+
+    if atoms.get_cell() is not  None:
+        cell1 = atoms.get_cell()[0]
+        cell2 = atoms.get_cell()[1]
+        cell3 = atoms.get_cell()[2]
+
+    #---------------STRUCT.fdf----------------
+    fileS = open(fname, 'w')
+    natm = len(atoms)
+    fileS.write("NumberOfAtoms    %d           # Number of atoms\n" % natm)
+    unique_symbs = get_unique_symbs(atoms)
+    fileS.write("NumberOfSpecies  %d           # Number of species\n\n" % len(unique_symbs))
+    fileS.write("%block ChemicalSpeciesLabel\n")
+
+    for symb in unique_symbs:
+        sym = ''.join(re.findall('[a-zA-Z]', symb))
+        fileS.write(" %d %d %s\n" % (unique_symbs.index(symb)+1,atomic_number(sym),symb) )
+    fileS.write("%endblock ChemicalSpeciesLabel\n")
+
+    #Lattice
+    fileS.write("\n#(3) Lattice, coordinates, k-sampling\n\n")
+    fileS.write("LatticeConstant   %15.9f Ang\n" % cellparameter)
+    fileS.write("%block LatticeVectors\n")
+    if atoms.get_cell() is not None:
+        va, vb, vc = cell1, cell2, cell3
+        fileS.write("%15.9f %15.9f %15.9f\n" % tuple(va))
+        fileS.write("%15.9f %15.9f %15.9f\n" % tuple(vb))
+        fileS.write("%15.9f %15.9f %15.9f\n" % tuple(vc))
+    fileS.write("%endblock LatticeVectors\n\n")
+
+    #Coordinates
+    fileS.write("AtomicCoordinatesFormat Ang\n")
+    fileS.write("%block AtomicCoordinatesAndAtomicSpecies\n")
+
+    for atom in atoms:
+        x,y,z = atom.get_position(); symb = atom.get_symbol()
+        if atom.get_groupid() != 0:
+            symb = symb + str(atom.get_groupid())
+        fileS.write(" %15.9f %15.9f %15.9f %4d %4d\n" %\
+                    (x,y,z,unique_symbs.index(symb)+1, atom.get_serial()))
+        
+    fileS.write("%endblock AtomicCoordinatesAndAtomicSpecies\n")
+    fileS.close()
+
+def read_struct(file_name):
+    vec_block = []; atoms_block = []; abc_cell_block = []
+    atoms_length = 0; species = []
+    n_of_species = 0; name = ''; atoms = []; cell = []; cell_scale = ''
+    lattice_constant = 0.
+    _is_ang_scale = 0; _is_bohr_scale = 0; _is_scaled_ang_scale = 0
+    _is_fraction_scale = 0
+
+    f = open(file_name)
+    lines = f.readlines()
+
+    i = 0
+    for line in lines:
+        #print i
+        
+        line_s = line.split(); keyword = ''
+
+        if line_s:
+            keyword = line_s[0].lower()
+            #print keyword
+
+        if keyword == "systemlabel":
+            name = line_s[1]
+
+        elif keyword == "latticeconstant":
+            lattice_constant = float(line_s[1])
+            #print lattice_constant
+            try:
+                cell_scale = line_s[2]
+            except:
+                cell_scale = 'Ang'
+
+        elif keyword == "atomiccoordinatesformat":
+            if line_s[1].lower() == 'ang':
+                _is_ang_scale = 1
+            elif line_s[1].lower() == 'bohr':
+                _is_bohr_scale = 1
+            elif line_s[1].lower() == 'scaledcartesian':
+                #print "ON"
+                _is_scaled_ang_scale = 1
+            elif line_s[1].lower() == 'fractional':
+                _is_fraction_scale = 1
+            else:
+                #print 'Warning : Default atomic scale, "Ang".\n'
+                pass
+
+        elif keyword == "numberofatoms":
+            atoms_length = int(line_s[1])
+            #print "natms", atoms_length
+
+        elif keyword == "numberofspecies":
+            n_of_species = int(line_s[1])
+            #print "nspec", n_of_species
+
+        elif keyword =="%block":
+            keyword_ = line_s[1].lower()
+            #print keyword_
+            
+            if keyword_ == "latticeparameters":
+                abc_cell_block = lines[i+1].split()
+
+            elif keyword_ == "latticevectors":
+                vec_block = lines[i+1:i+4]
+
+            elif keyword_ == "atomiccoordinatesandatomicspecies":
+                atoms_block = lines[i+1:i+1+atoms_length]
+                #print "atoms_block", atoms_block
+
+            elif keyword_ == "chemicalspecieslabel":
+                temp = lines[i+1:i+1+n_of_species]
+                for spec in temp:
+                    species.append(spec.split()[2])
+                #print species
+        i +=1
+
+    # cell converting
+    va = 0; vb = 0; vc = 0
+    if (not abc_cell_block) and vec_block:
+        a1, a2, a3 = vec_block[0].split()
+        a1 = float(a1); a2 = float(a2); a3 = float(a3)
+        b1, b2, b3 = vec_block[1].split()
+        b1 = float(b1); b2 = float(b2); b3 = float(b3)
+        c1, c2, c3 = vec_block[2].split()
+        c1 = float(c1); c2 = float(c2); c3 = float(c3)
+        va = np.array([a1, a2, a3])
+        vb = np.array([b1, b2, b3])
+        vc = np.array([c1, c2, c3])
+        if cell_scale == 'Ang':
+            va = lattice_constant * va
+            vb = lattice_constant * vb
+            vc = lattice_constant * vc
+        elif cell_scale == 'Bohr':
+            va = lattice_constant * bohr2ang * va
+            vb = lattice_constant * bohr2ang * vb
+            vc = lattice_constant * bohr2ang * vc
+        else:
+            #print "Can`t find cell scale"
+            pass
+
+        #a, b, c, alpha, beta, gamma = convert_xyz2abc(va, vb, vc)
+        cell = np.array([va,vb,vc])
+
+    elif abc_cell_block and (not vec_block):
+        a, b, c, alpha, beta, gamma = abc_cell_block.split()
+        a = float(a); b = float(b); c = float(c)
+        alpha = float(alpha); beta = float(beta); gamma = float(gamma)
+        cell = [a, b, c, alpha, beta, gamma]
+
+    # atoms
+    for atm in atoms_block:
+
+        if len(atm.split()) == 4:
+            x, y, z, spec = atm.split()
+        elif len(atm.split()) == 5:
+            x, y, z, spec, serial = atm.split()
+
+        x = float(x); y = float(y); z = float(z); spec = int(spec)
+
+        if _is_ang_scale:
+            pass
+        elif _is_bohr_scale:
+            x = bohr2ang * x; y = bohr2ang * y; z = bohr2ang * z
+
+        elif _is_scaled_ang_scale:
+            #if vec_cell:
+            x = lattice_constant*x
+            y = lattice_constant*y
+            z = lattice_constant*z
+            #elif not vec_cell:
+            #    print "Can`t guess cell scale and type\n"
+#        elif _is_fraction_scale:
+        
+        group = re.findall('\d', species[spec-1])
+        sym = ''.join(re.findall('[a-zA-Z]', species[spec-1]))
+        if group:
+            group = int(''.join(group))
+        else:
+            group = None
+        
+        atom = Atom(sym, (x, y, z), groupid=group)
+        atoms.append(atom)
+
+    if cell.shape == (3,3):
+        #XYZ.write_xyz(file_name.replace('fdf','xyz'), atoms, cell)
+        return AtomsSystem(atoms, cell=cell)
+    else:
+        #XYZ.write_xyz(file_name.replace('fdf','xyz'), atoms)
+        return AtomsSystem(atoms, cell=None)
+
 
 #
 # REload saved simulation
