+++ units/unitconvert.py
@@ -2,6 +2,7 @@ from math import pi
 
 ### unit conversions
 
-ang2bohr = 1.889726878      # Angstrom  to bohr     conversion
-rad2deg = 180/pi            # radian    to degree   conversion
-degrad = 1./rad2deg
+ang2bohr    = 1.889726878      # Angstrom  to bohr     conversion
+rad2deg     = 180/pi            # radian    to degree   conversion
+degrad      = 1./rad2deg
+bohr2ang    = 1./ang2bohr
